remove support for ExtUtils::ParseXS argtypes flag
If the ExtUtils::ParseXS argtypes flag is set, function declarations
are parsed as ANSI-C-like declarations, otherwise they are parsed as
old-style declarations. It defaults to being set, and ExtUtils::MakeMaker
doesn't tweak it, so I'm not sure if the non-ANSI code ever gets used.

In any case, the code here should be robust enough to handle either
style. Removing the extra code to reduce complexity and make sure all
arguments are parsed the same way.